requirements.txt
streamlit
pypdf2
langchain
python-dotenv
faiss-cpu
openai
InstructorEmbedding
tiktoken
# huggingface_hub
# sentence_transformers